requirements.txt
django==1.11a1
coreapi==2.1.1
djangorestframework==3.5.3